moved buttons and animal into animalview
